Add proper support for updating hardware
This adds some tests for the PATCH path, which never was under
test and had some bugs. Had to move some validator logic around a bit
due to incompat b/w the validators and the way flask wants to put
path params on the request.

The hardware PATCH endpoint also should allow for availability windows
to be added/removed/updated, but there are not yet tests for this,
so who knows if it really works ;)