Add WorkerField to encapsulate field logic
This adds a WorkerField abstraction, which aims to solve a few problems:

- Workers may depend or require some hardware property fields.
- Some of these fields may be sensitive and should never be rendered
- Some of the fields are only for the device owners
- Workers can be enabled or disabled, without updating the hardware
  properties in the database.

The worker fields allow validation on the incoming field values and also
allow filtering the output to e.g., drop fields which are not known to
any worker.

This does mean that any field MUST be declared by a worker, otherwise it
will be hidden. This only applies to fields w/in the "properties" structure
for hardware objects.