Add start of Ironic plugin
BREAKING: adds "context" argument to worker processing.

This adds a bulk of the logic for the Ironic plugin. It looks like
there is a need to have workers surface opts to the service; that
will come in a future commit. There are also more tests to be written.

There is a new argument in the function signature for the worker
process function; a context is needed to do most everything and it makes
sense to pass this to the worker at processing time.