Use names for foreign key constraints
See https://alembic.sqlalchemy.org/en/latest/naming.html for a whole
lot of information about why this is needed. We use declarative base
style of models so I'm not sure how to leverage MetaData for automatic
naming conventions. Lesson learned, always name the constraints.

Without a name, the first arg in the constraint migrations is None, which
is not valid.