Fix bug where properties not validated
If properties were not sent, they were not validated b/c this field
was not required. Add a require on it; in practice we'll always be
using properties for something so it should always be sent by the client
on enrollment.

Also fixes issue with management address schema and requires at least
one interface be defined.