Feature/blazar sync worker (#8)
This PR adds the Blazar Sync worker, for Blazar Hosts and Leases.

Host creation and update:

To avoid expensive lookups, it caches the mapping from blazar_host_id to hardware_uuid in the state_details of the task. Otherwise, we can only discover this mapping by listing all hosts, then searching for the matching hardware_uuid.

If the blazar_host_id is not present, we assume that the host must be created. In the case that the host already exists, the worker will search for the host, and add its ID to the cache.

On Duplicate detection:
We assume that hardware_uuid is unique, and use it to detect several failure cases in blazar's logic.
The name field is used by Blazar to check the status of the host in nova, and thus ironic. Our convention is to have name == hardware_uuid, so we continue that here. Blazar checks for 3 cases: Missing, Present, and Multiple.

In the "Missing" case, we defer, as the Ironic worker must complete before this one.
The "Present" case is the "base" case, and host creation proceeds normally.
The "Multiple" case is an error state, but ironic should usually prevent this from occurring.

If the blazar_host_id is present, we assume the "update" case.
We unconditionally update, rather than GET + UPDATE, to reduce the number of calls.

If the host cannot be updated due to a conflict, we detect a 409 and defer.
If the host cannot be found, due to an invalid host_id, we first search for it, and store either the correct host ID, or set it to NULL to rerun CREATE.

Lease creation, update, and deletion:

Each host in doni has a list of AvailabilityWindow associated with it. This worker maps each one to a lease in Blazar, named "availability_window_" followed by the UUID of the AvailabilityWindow
This is in order to distinguish these leases from those created manually, or by other tasks.

On every run, the worker GETs the list of all leases from Blazar, and then filters it for those with the above prefix. Then, for each window in the list passed to the worker, it CREATEs it if not found in blazar, UPDATEs if the window's properties are not a subset of the existing lease, and ignores it otherwise.

At the end, any leases that match the prefix, and have not been found in the list of windows are DELETEd.