Handle shutdown signal during startup
If a shutdown happened when the manager was in the middle of
initialization, an attribute error could be raised b/c some attributes
were never initialized to a value.

This also may fix an issue where the periodic worker tasks were never
waited upon; perhaps this is the root of some tasks being stuck in
progress erroneously.